Adicionando validações e corrigindo códigos de retorno em UserController.

+++ src/main/java/io/github/lucasduete/pweb2/serviceauth/controllers/UserController.java
@@ -7,6 +7,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
+
 @RestController
 @RequestMapping("user")
 public class UserController {
@@ -20,7 +22,7 @@ public class UserController {
         if (user == null)
             return ResponseEntity
                     .status(HttpStatus.BAD_REQUEST)
-                    .body("Usuário não foi informado");
+                    .body(null);
 
         return ResponseEntity.ok(this.userService.save(user));
     }
@@ -28,38 +30,68 @@ public class UserController {
     @GetMapping
     public ResponseEntity getAll() {
 
-        return ResponseEntity.ok(
-                this.userService.listAll()
-        );
+        List<User> users = this.userService.listAll();
+
+        if (users == null || users.isEmpty())
+            return ResponseEntity
+                    .status(HttpStatus.NO_CONTENT)
+                    .body(null);
+
+
+        return ResponseEntity.ok(users);
     }
 
     @GetMapping("{id}")
     public ResponseEntity getById(@PathVariable Long id) {
+
+        if (id <= 0)
+            return ResponseEntity
+                    .status(HttpStatus.BAD_REQUEST)
+                    .body(null);
 
         User user = this.userService.getById(id);
 
         if (user == null)
             return ResponseEntity
                     .status(HttpStatus.NOT_FOUND)
-                    .body("Este usuário não exite");
+                    .body(null);
 
         return ResponseEntity.ok(user);
     }
 
     @PutMapping("{id}")
-    public User updateUser(@RequestBody User user, @PathVariable Long id) {
+    public ResponseEntity updateUser(@RequestBody User user, @PathVariable Long id) {
+
+        if (id <= 0)
+            return ResponseEntity
+                    .status(HttpStatus.BAD_REQUEST)
+                    .body(null);
+
         User userById = this.userService.getById(id);
+
+        if (userById == null)
+            return ResponseEntity
+                    .status(HttpStatus.BAD_REQUEST)
+                    .body(null);
 
         userById.setMatricula(user.getMatricula());
         userById.setPassword(user.getPassword());
         user.setRole(user.getRole());
 
-        return this.userService.save(userById);
+        return ResponseEntity.ok(this.userService.save(userById));
     }
 
     @DeleteMapping("{id}")
-    public void deleteUser(@PathVariable Long id) {
+    public ResponseEntity deleteUser(@PathVariable Long id) {
+
+        if (id <= 0) {
+            return ResponseEntity
+                    .status(HttpStatus.BAD_REQUEST)
+                    .body(null);
+        }
+
         userService.deleteById(id);
+        return ResponseEntity.ok(null);
     }
 
 }
